% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-sma.R, R/stat-sma.R
\name{geom_sma}
\alias{geom_sma}
\alias{stat_sma}
\title{Smoothed conditional means}
\usage{
geom_sma(
  mapping = NULL,
  data = NULL,
  stat = "sma",
  position = "identity",
  ...,
  method = NULL,
  formula = NULL,
  se = TRUE,
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE
)

stat_sma(
  mapping = NULL,
  data = NULL,
  position = "identity",
  ...,
  method = "sma",
  formula = NULL,
  se = TRUE,
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{geom, stat}{Use to override the default connection between
`geom_smooth()` and `stat_smooth()`.}
}
\description{
Aids the eye in seeing patterns in the presence of overplotting.
`geom_smooth()` and `stat_smooth()` are effectively aliases: they
both use the same arguments. Use `stat_smooth()` if you want to
display the results with a non-standard geom.
}
\details{
Calculation is performed by the (currently undocumented)
`predictdf()` generic and its methods.  For most methods the standard
error bounds are computed using the [predict()] method -- the
exceptions are `loess()`, which uses a t-based approximation, and
`glm()`, where the normal confidence interval is constructed on the link
scale and then back-transformed to the response scale.
}
\examples{
ggplot(mpg, aes(displ, hwy)) +
  geom_point() +
  geom_sma()
}
\seealso{
See individual modelling functions for more details:
  [lm()] for linear smooths,
  [glm()] for generalised linear smooths, and
  [loess()] for local smooths.
}
