% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsma.R
\docType{package}
\name{ggsma-package}
\alias{ggsma}
\alias{ggsma-package}
\title{SMA lines}
\description{
Functions for obtaining the density, random deviates and maximum likelihood
estimates of the zero-truncated Poisson lognormal distributions and their
mixtures.
}
\references{
Bulmer, M. G. 1974. On Fitting the Poisson Lognormal Distribution to Species-Abundance Data. Biometrics 30:101-110.
}
\author{
Masatoshi Katabuchi <mattocci27@gmail.com>
}
\keyword{internal}
