% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsma.R
\docType{package}
\name{ggsma-package}
\alias{ggsma}
\alias{ggsma-package}
\title{SMA lines}
\description{
Functions for obtaining the density, random deviates and maximum likelihood
estimates of the zero-truncated Poisson lognormal distributions and their
mixtures.
}
\references{
Warton, David I., Ian J. Wright, Daniel S. Falster, and Mark Westoby. 2006. Bivariate Line-Fitting Methods for Allometry. Biological Reviews 81: 259–91.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mattocci27/ggsam}
  \item Report bugs at \url{https://github.com/mattocci27/ggsam/issues}
}

}
\author{
Masatoshi Katabuchi <mattocci27@gmail.com>
}
\keyword{internal}
